moved functions to controller.py and also uncommented ACL part
